utils: move Profile Wizard code to new file
In preparation for the addition of the AD9528 parser, move Profile
Wizard code to it's own file.
