+++ adidt/parts/adrv9009.py
@@ -1,6 +1,7 @@
 import xmltodict
 from typing import Dict
 from adidt.dt import dt
+from adidt.utils import profilewiz
 import fdt
 import math
 
@@ -11,92 +12,8 @@ def handle_ints(val):
     return int(hex((val + (1 << 32)) % (1 << 32)), 16)
 
 
-def coefs_to_long_string(coefs):
-    """Convert coefficient array to string.
-
-    Args:
-        coefs (list): Coefficients.
-
-    Returns:
-        str: Coefficients as a string.
-    """
-    result = ""
-    for coef in coefs.split("\n"):
-        coef = coef.replace(" ", "")
-        result += f"({coef}) "
-    return result[:-1]
-
-
-def profile_to_xml(filename):
-    # Update profile to non-shitty xml
-    with open(filename) as sxmlfile:
-        sxmldata = sxmlfile.read()
-    # Correct each header
-    outfile = []
-    for line in sxmldata.split("\n"):
-        if "<" in line and "</" not in line:
-            within = line[line.find("<") + 1 : line.find(">")]
-            # Fix all prop=val
-            if "," in within:
-                oline = []
-                # print(line)
-                for index, sline in enumerate(line.split(" ")):
-                    if index == 3:
-                        sline = f"Bandwidth={sline}"
-                    if index in [4, 6]:
-                        sline = sline + "="
-                    oline += [sline]
-                line = " ".join(oline)
-                line = line.replace(",", "")
-                line = line.replace("= ", "=")
-            if " " in within:
-                # print(line)
-                oline = []
-                spaces = 0
-                for sline in line.split(" "):
-                    if len(sline) > 0:
-                        starter = sline + " "
-                        break
-                    spaces += 1
-                for sline in line.split(" "):
-                    if (
-                        len(oline) == 0
-                        and len(sline)
-                        and "<" not in sline
-                        and ">" not in sline
-                        and "=" not in sline
-                    ):
-                        oline += [f'type="{sline}"']
-                    if "=" in sline:
-                        sline = sline.replace(">", "")
-                        sline = sline.replace("<", "")
-                        p = sline.split("=")[0]
-                        v = sline.split("=")[1]
-                        # print(f"{p}={v}")
-                        oline += [f'{p}="{v}"']
-
-                # print(" "*spaces+starter+" ".join(oline)+">")
-                outfile += [" " * spaces + starter + " ".join(oline) + ">"]
-            else:
-                # print(line)
-                if "=" in line:
-                    s1 = line.find("<")
-                    s2 = line.find(">")
-                    within = line[s1 + 1 : s2]
-                    o = within.split("=")
-                    ss = f"{o[0]}>{o[1]}</{o[0]}"
-                    line = line.replace(within, ss)
-                # print(line)
-
-                outfile += [line]
-        else:
-            outfile += [line]
-
-    return "\n".join(outfile)
-
-
 def parse_profile(filename):
-    nsxml = profile_to_xml(filename)
+    nsxml = profilewiz.profile_to_xml(filename)
     profile = xmltodict.parse(nsxml)["profile"]
 
     rx = profile['rx']
@@ -160,22 +77,22 @@ def parse_profile(filename):
 
         tx["filter"]["@gain_dB"] = f"({tx['filter']['@gain_dB']})"
 
-    rx["rxAdcProfile"]["coefs"] = coefs_to_long_string(rx["rxAdcProfile"]["#text"])
+    rx["rxAdcProfile"]["coefs"] = profilewiz.coefs_to_long_string(rx["rxAdcProfile"]["#text"])
     del(rx["rxAdcProfile"]["#text"])
-    rx["filter"]["coefs"] = coefs_to_long_string(rx["filter"]["#text"])
+    rx["filter"]["coefs"] = profilewiz.coefs_to_long_string(rx["filter"]["#text"])
     del(rx["filter"]["#text"])
 
-    orx["filter"]["coefs"] = coefs_to_long_string(orx["filter"]["#text"])
+    orx["filter"]["coefs"] = profilewiz.coefs_to_long_string(orx["filter"]["#text"])
     del(orx["filter"]["#text"])
-    orx["orxBandPassAdcProfile"]["coefs"] = coefs_to_long_string(orx["orxBandPassAdcProfile"]["#text"])
+    orx["orxBandPassAdcProfile"]["coefs"] = profilewiz.coefs_to_long_string(orx["orxBandPassAdcProfile"]["#text"])
     del(orx["orxBandPassAdcProfile"]["#text"])
-    orx["orxLowPassAdcProfile"]["coefs"] = coefs_to_long_string(orx["orxLowPassAdcProfile"]["#text"])
+    orx["orxLowPassAdcProfile"]["coefs"] = profilewiz.coefs_to_long_string(orx["orxLowPassAdcProfile"]["#text"])
     del(orx["orxLowPassAdcProfile"]["#text"])
 
-    tx["filter"]["coefs"] = coefs_to_long_string(tx["filter"]["#text"])
+    tx["filter"]["coefs"] = profilewiz.coefs_to_long_string(tx["filter"]["#text"])
     del(tx["filter"]["#text"])
 
-    lpbk["lpbkAdcProfile"]["coefs"] = coefs_to_long_string(lpbk["lpbkAdcProfile"]["#text"])
+    lpbk["lpbkAdcProfile"]["coefs"] = profilewiz.coefs_to_long_string(lpbk["lpbkAdcProfile"]["#text"])
     del(lpbk["lpbkAdcProfile"]["#text"])
 
     return {"rx": rx, "tx": tx, "orx": orx, "lpbk": lpbk, "clocks": clocks}
